Cleanup the BM(P|E) sensors a bit
Limit the scope of variables. By doing so we don't create variables
that are initialy uninitialized. A limited scope makes code easier
to read (the reader only needs to consider a small amount of code).

Only overwrite the last_value_(...) variables when actually on the
error path. Is not needed on the general path.

Don't initialise the String with "". The string class does this
itself.

Try to cleanup some confusion around the types used for the measurements.
readPressure() and friends all return floats. We store them in doubles
and sometimes ints? The last_value_(...) variables are again doubles.
I just leave the measurements as floats (by using auto). last_value_(...)
stays a double for now.